[RU] Subtitles for Chapter 1 of the video course (#489)
* Created a directory for the russian subtitles.

Created a folder for Russian subtitles for the video course and published a translation of the introductory video from chapter 1.

* Uploaded subtitles for chapter 1

Uploaded subtitles for the remaining videos for chapter 1 of the video course.

* Added subtitles for chapter 2 of the video course

Added STR subtitle files for the second chapter of the YouTube video course.

* Delete subtitles/ru directory

Removed the old translation. Incorrect timestamping.

* Create 00_welcome-to-the-hugging-face-course.srt

Create a directory and upload a subtitle file for the introductory video of the course.

* Add files via upload

Upload subtitle files for the first chapter of the course.